data_dictionary_AFINN {quanteda.dictionaries}	R Documentation

Affective Norms for English Words (AFINN)

Description

A quanteda dictionary object containing the Affective Norms for English Words (AFINN), a publicly available list of English words rated for valence with values between -5 (negative) and +5 (positive). AFINN-111, the latest version, contains 2,477 words and phrases. Note that the version implemented in quanteda.dictionaries uses a binary classification into 878 positive (valence > 0) and 1,599 negative (valence < 0) features.

Usage

data_dictionary_AFINN

Format

An object of class dictionary2 of length 2.

Source

http://www2.imm.dtu.dk/pubdb/views/publication_details.php?id=6010

References

Nielsen, Finn Å. (2011). "A new ANEW: Evaluation of a Word List for Sentiment Analysis in Microblogs." In Proceedings of the ESWC2011 Workshop on 'Making Sense of Microposts': Big Things Come in Small Packages: 93-98.

[Package quanteda.dictionaries version 0.2 Index]