data_dictionary_HuLiu {quanteda.dictionaries}	R Documentation

Positive and negative words from Hu and Liu (2004)

Description

A quanteda dictionary object containing 2,006 positive and 4,783 negative words from Hu and Liu (2004, 2005).

Usage

data_dictionary_HuLiu

Format

An object of class dictionary2 of length 2.

Source

http://www.cs.uic.edu/~liub/FBS/sentiment-analysis.html

References

Minqing Hu and Bing Liu. "Mining and Summarizing Customer Reviews." In Proceedings of the ACM SIGKDD International Conference on Knowledge Discovery and Data Mining (KDD-2004), Aug 22-25, 2004, Seattle, Washington, USA.

Bing Liu, Minqing Hu and Junsheng Cheng. "Opinion Observer: Analyzing and Comparing Opinions on the Web." In Proceedings of the 14th International World Wide Web conference (WWW-2005), May 10-14, 2005, Chiba, Japan.

[Package quanteda.dictionaries version 0.2 Index]