data_dictionary_LaverGarry {quanteda.dictionaries}	R Documentation

Laver and Garry Dictionary of Policy Positions

Description

A quanteda dictionary object containing the Laver and Garry dictionary which been developed to estimates the policy positions of political actors in the United Kingdom by comparing their speeches and written documents to key words found in the British Conservative and Labour manifestos of 1992. Note: Please remember that this dictionary was customized to reflect the policy positions of UK political parties.

Usage

data_dictionary_LaverGarry

Format

The dictionary has two levels of nesting with 7 main policy areas (level 1) divided up into 19 sub-categories (level 2).

Source

https://provalisresearch.com/products/content-analysis-software/wordstat-dictionary/laver-garry-dictionary-of-policy-position/

References

Laver. M., and Garry, J. (2000). "Estimating Policy Positions from Political Texts." American Journal of Political Science 44(3): 619-634.

[Package quanteda.dictionaries version 0.2 Index]