data_dictionary_MFD {quanteda.dictionaries}	R Documentation

Moral Foundations Dictionary

Description

A quanteda dictionary object containing the Moral Foundations Dictionary, a publicly available dictionaries with information on the proportions of virtue and vice words for each foundation. The categories are harm (vice/virtue), fairness (vice/virtue), ingroup (vice/virtue), authority (vice/virtue), purity (vice/virtue) and morality (general).

Usage

data_dictionary_MFD

Format

An object of class dictionary2 of length 11.

Source

http://moralfoundations.org/othermaterials

References

Haidt, J., Graham, J., and Nosek, B.A. (2009). "Liberals and Conservatives Rely on Different Sets of Moral Foundations. Journal of Personality and Social Inquiry 20(2-3): 110-119.

Graham, J., and Haidt, J. (2016). Moral Foundations Dictionary.: http://moralfoundations.org/othermaterials.

[Package quanteda.dictionaries version 0.2 Index]