data_dictionary_NRC {quanteda.dictionaries}	R Documentation

NRC Word-Emotion Association Lexicon

Description

A quanteda dictionary object containing Mohammad and Charron's (2010, 2013) English version of the NRC Word-Emotion Association Lexicon (aka NRC Emotion Lexicon aka EmoLex): association of words with eight emotions (anger, fear, anticipation, trust, surprise, sadness, joy, and disgust) and two sentiments (negative and positive) manually annotated on Amazon's Mechanical Turk. Available in 40 different languages.

Usage

data_dictionary_NRC

Format

An object of class dictionary2 of length 10.

References

Saif Mohammad and Peter Turney (2013). "Crowdsourcing a Word-Emotion Association Lexicon." Computational Intelligence 29(3), 436-465.

Saif Mohammad and Peter Turney (2010). "Emotions Evoked by Common Words and Phrases: Using Mechanical Turk to Create an Emotion Lexicon." In Proceedings of the NAACL-HLT 2010 Workshop on Computational Approaches to Analysis and Generation of Emotion in Text, June 2010, LA, California.

NRC webpage: http://saifmohammad.com/WebPages/AccessResource.htm

[Package quanteda.dictionaries version 0.2 Index]