data_dictionary_geninqposneg {quanteda.dictionaries}	R Documentation

Augmented General Inquirer Positiv and Negativ dictionary

Description

A quanteda dictionary object containing the Positiv and Negativ dictionary entries from the augmented General Inquirer. These are new valence categories described at http://www.wjh.harvard.edu/~inquirer/homecat.htm but also including the terms from the yes and no dictionary entries.

Usage

data_dictionary_geninqposneg

Format

An object of class dictionary2 of length 2.

Source

http://www.wjh.harvard.edu/~inquirer/spreadsheet_guide.htm

References

Stone, Philip J., Dexter C. Dunphy, and Marshall S. Smith. 1966. The General Inquirer: A Computer Approach to Content Analysis. Cambridge, MA: MIT Press.

[Package quanteda.dictionaries version 0.2 Index]