data_dictionary_sentiws {quanteda.dictionaries}	R Documentation

SentimentWortschatz (SentiWS)

Description

A quanteda dictionary object containing SentimentWortschatz (SentiWS), a publicly available German-language resource for sentiment analysis. The current version of SentiWS contains 1,650 positive and 1,818 negative words, which sum up to 15,649 positive and 15,632 negative word forms including their inflections. It not only contains adjectives and adverbs explicitly expressing a sentiment, but also nouns and verbs implicitly containing one. The original dictionary weights within the interval of -1 to 1. Note that the version implemented in quanteda.dictionaries uses a binary classification into positive (weight > 0) and negative (weight < 0) features.

Usage

data_dictionary_sentiws

Format

An object of class dictionary2 of length 2.

Source

http://wortschatz.uni-leipzig.de/en/download/

References

Remus, R., Quasthoff U., and Heyer, G. (2010). "SentiWS - a Publicly Available German-language Resource for Sentiment Analysis." In Proceedings of the 7th International Language Ressources and Evaluation (LREC'10): 1168-1171.

[Package quanteda.dictionaries version 0.2 Index]